COMMERCE-API Move class in the right package as standard
